It lives!
Now we are able to plot things. This needs a good review of the values
produced, since most likely there will be bugs (this is all done in a
hurry, and it's just hard to test manually, and probably even more
automatically). But so far I'm getting values which look meaningful.