Try a new approach without patching Qt
Could be that my Qt patch, that I expected to be the solution for the
axes not updating correctly after repeated calls to createDefaultAxes,
will not be accepted, because one is supposed to set the range manually.
Seems an inconvenient API, but I won't be able to do anything better for
now.