Refactor: don't assume that much on QLineSeries
Right now the chart only shows N line series. But I'm thinking how to
improve visualization by showing things like a detailed stacked bar
chart with the breakdown of where the damage comes (weapon, strength,
proficiency, etc.) as well as an indication of how much some
debilitating effect of the weapon (poison, stun, etc.) appears.

So this refactors the code to not assume that only line series that
match the number of tabs are the only series of the chart.

This still misses a check for memory leaks, just in case, but valgrind
is failing on my laptop.