Add Backstab Calculator page
It's mostly copied from the Damage Calculator one, but with a completely
new UI on the side for a different kind of input. Not showing any chart
yet, as the calculation has to be decided a bit where to put it, so it's
testable and well designed from the beginning, as I don't want to screw
up like with Damage Calculator.