Fully transition DC's UI from monolithic to modular
Break out the damagecalculationwidget.ui into little parts without
repetition. Each weapon is now an instance of WeaponArrangementWidget,
with the differences per weapon sorted out later. The different APR kind
(double vs int combo box) is sorted by having two, and hiding the
inappropiate one. The different ranges, by just changing them for each
one.

There is still a lot of uglyness to be cleaned up, but it would probably
require moving the serialization functions to a testable location, which
would require moving the widgets to the library as well, and start
depending on UI things there and in the tests.