Ugly workaround to layout issues
Each widget added to a QToolBox is added to a QScrollArea. Since we have
tool boxes inside another tool box, when there is not enough vertical
space, the scrollbars appear in the innermost scroll area. Add a minimum
size to an inner group box, so the inner tool box inside always has room
and doesn't require scrollbars. Then the outer one shows them when
needed.

This is far from polished, and I feel that it's wrong, but it's the best
I can think of right now for a quick solution given my usual lack of
patience with layouts.