Add a lot of new features to DiceRoll
While implementing the new damage calculator features (supporting
elemental damage) I realized that I did not calculate resistance
properly: it's not enough calculating average damage normally, then
applying resistance. The resistance is applied to each roll
individually, so a 2d4+2 roll would almost not be affected at 10%
resistance, because only the highest 10 gets reduced by 1 (the rest, get
the damage rounded up).

To calculate the standard deviation, a new function has been added to
return the whole set of possible permutations, so we can add each of
them appropriately to determine the effect of resistance.

Add a helper `resistified` akin to an existing `luckified` and test it.
Change the implementation of most functions to go through it, so we
supposedly(!) have everything well covered.

Add a `probability` feature to consider the impact on average and
variance of the non-100% probability of the roll being made (so 0 is
assumed). This is the only part that needs more unit testing.

Also add a comparison operator and relax the assertion that checks the
lowest possible value for number, because the damage calculator is going
to need it right away.