DiceRoll constructor improvements
Add a new constructor based on a struct to simulate named parameters,
inspired by: https://pdimov.github.io/blog/2020/09/07/named-parameters-in-c20/

Reorder members for consistency with the "1d8+2" notation.

Skip some parts on MSVC, as the use of designated initializers here is
based on GCC/Clang extensions, not on C++ 20 yet.