Fix the menus, specially on WASM
I had an unconditionally called setEnabled(false), which I did not
notice at all, given that the Plasma global menu doesn't work too well
when changing the state of things. Remove it, and disable instead the
specific menu entries on WASM, because we would need to switch to an
asynchronous way of polling QSettings for readyness, which sucks.