Improvements for load calculations from file
Clean the existing dialog code a bit to avoid std::bind, use deleteLater
(dialog was not leaking, but still), and set a name filter.

Add the WASM dialog version that works differently, and adjust the
loading function for the dialog limitations under WASM. Left a debug
trace in place, because I'm not very confident in the documentation of
the new QFileDialog::getOpenFileContent function, specially after seeing
the example (where it seems to imply that the file name could be empty).