Try a QSplitter in the Damage Calculator
On my Windows, at least, the defaults are pretty stupid. I can't get the
input controls to take as much space as needed and not show the
scrollbar. Given how much this sucks, I'll switch to a splitter, so at
least users can change it to their needs.

I will try to make it better with GammaRay on Linux.