A much better splitter (hopefully) in Damage Calc
The damn QToolBox class returns a sizeHint value that doesn't seem to
make sense when the widget is put inside the splitter. When I had the
widget on the horizontal layout, the tool box is returning about 530 px,
which is the smallest width that the biggest widget inside it requires
to be shown without a hideous horizontal scroll bar. So that value makes
sense.

However, as soon as I put the inputArea QToolBox into the QSplitter, the
size hint of the former shrinks into about 80 px. After some failed
attempts at setting different size policies, I gave up and just went to
subclassing the QToolBox, which is dead easy and works. Now it returns
the maximum value of the children's size hints in its size hint.

Also, setting the sizes of the splitter children on the constructor
sucks, as the widgets are not visible at that time, so moving that to a
resize event or the like is needed. Completed this work by saving and
loading the state of the splitter to the settings, which paves the way
to maybe keeping the whole MainWindow state as well.