Change the QKeySequences to avoid enum adding
With the latest C++ version enabled in the build, things like
Qt::CTRL+Qt::Key_S produce a warning because it's adding two different
enums. Since wrapping them in integers is a bit ugly, and maybe error
prone if the underlying integral type changes, just concede the use of
standard shortcuts where possible, and strings otherwise (which I
normally avoid, but that here at least have the advantage of being
translatable if needed).