New strategy for menus
Add a function to the BasePage class for giving pages the opportunity to
create menus themselves, yet allow the main window to hide and show
them.

Added a menu to the backstab calculator to copy the chart to clipboard
using this mechanism.

Should pave the way for improving the menus on the damage calculator,
which are going to require a lot of work still.