Change the text on the Welcome page
Even though at this point I have avoided giving the application to
anyone before the "full" 0.2 release and before I finish some
documentation, one day I gave the link to the online WASM version to one
person. It did not go too well, and that person seemed to got stuck in
the first section of the welcome page, not knowing what to do.

I think the text was properly written as it was, and it says that you
don't need to use that section and can skip to press the buttons below.
And I instructed that person to do so in the messages where I gave the
URL. But still... I worry that it's not noob-friendly enough.

So try to restructure the description this time with a bit more markup,
and using bullet points on how to actually setup a new game.

While we are at it, add the ellipsis to the "Change..." button that
opens a dialog, as per the usual HIG convention.