Add and test a new function for hitting or not
Instead of past approaches of "chance to hit" directly calculated in a
floating point number like a %, I want now to calculate if each roll of
a given configuration would fail, hit, or produce a critical hit. So
by returning 0, 1 or 2, I'll be able to multiply it by the damage on
hit, and hence calculate better the average damage for the 1-20 range.