Rework damage calculations
The Damage Calculator had an embarrassing mistake in how it calculated
the critical hit damage. It was plain wrong. Since fixing it entailed
thinking a bit ahead of time to consider the feature of supporting a
variable amount of critical misses in addition to the variable critical
hits, and that part of the code was pending to be moved to a testable
class, do all that at once.

Calculate the average damage by calculating first the distribution of
misses, hits and critical hits in the 1 to 20 range. Then use that and a
better API for the regular hits and critical misses to calculate the
average damage. We can still make it even more testable by moving the
multiplication of APR, damage on hit, etc. That should be in the
calculator class. But it's so simple that it doesn't matter much.

Now we support the critical hit chance, which is something I've seen
affected by mods (e.g. a certain Kensai mod adds the ability to never
make a critical miss). We are now also well prepared to be able to make
a "breakdown" second chart (as a tooltip, maybe) that can give the
details of the damage at a certain point.