Embarrassing fix: AC modifiers were wrong
I've been treating the AC modifiers as bonuses to AC in the calculation
(so, a positive input is making it a better AC, and a negative one
worse), but the values that were setup as defaults in the selector were
wrong, with the sign reversed to what they should have been (e.g. Full
Plate should have positive bonuses to make AC lower).

Keep the logic in the code as it was, but make it clearer in the UI by
showing "bonus" instead of "modifier". Fix it by swapping the sign of
all of them in the UI, and fixing the values for Leather armor, which
was also mistaken...