Redo GameBrowserPage with a filterable table
Drop the tree idea. It's only reason to be is to have the items nicely
folded and showing less initially, but I don't think it's actually a
good idea in the end. It was more "wishful thinking" on my part,
wanting to consider a tree as more powerful and with more possibilities,
but I no longer think it makes sense. The filterable table is still
limited, but already looks more complete and useful than before. A tree
with only one level of depth doesn't make sense.