Small ResourceManager/GameBrowserPage improvemnts
Add a few TODO notes for the next refactor of the ResourceManager. It
will need a few improvements to support the listing of override files in
the GameBrowserPage, and address older thoughts that were also written
down in the comments.

For now remove the standalone function in the resourcetype.h header, and
place the functionality in the ResourceManager as a static helper
function; which is pretty much an existing function but repurposed.