Don't pass TOML_HEADER_ONLY to every TU
In a recent refactor, a custom tomlplusplus.h file was created that sets
the proper macros for tomlplusplus and includes the "ugly" path for the
library. Every file that needs to use the library should use that, and
hence not need TOML_HEADER_ONLY set in every translation unit, which was
set "just in case only" anyway.