Big improvements to the Game Browser filter
Now it can support "smart" queries like "imoen is:cre in:data" or
"is:2da is:ids" to search for resources by name and by some of the other
columns which are supported right now.

Since this functionality is highly app specific, and I considered the
helper code to end up requiring GUI classes, this commit introduces
"libtoolkit", an app specific library that is just needed as an
implementation detail for easy testing.