Add two simple helper functions for path
I end up rewriting this in every project... At least make it shared and
minimally tested for a time.

It's weird to create this functions when QFileInfo exists, indeed, but I
had a too painful experience with some classes that sometimes did go
into the file system to do simple string manipulations that could avoid
the disk entirely.

This functions are surely not needed in most cases for most people,
though, as QFileInfo should do fine. It's a bit of a pet peeve of mine,
and to have fun.