Better display of locationError
Make it always visible. Just clearing its text is enough in Qt 6 to make
it display nicely with an empty text without the rest of the widgets
"dancing" up or down (Qt 5 seems to have a bug with the sizeHint of
QLabel that errs by 1 pixel, so forget about that).